fix crash during db lookup on 64 bit platforms

+++ dbprovider.c
@@ -19,7 +19,7 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
  *
- * $Id: dbprovider.c,v 1.23 2009/08/13 20:09:20 mdpetch Exp $
+ * $Id: dbprovider.c,v 1.24 2009/08/14 19:07:42 c_anthon Exp $
  */
 
 #include "config.h"
@@ -96,8 +96,8 @@ static RowSet* MallocRowset(size_t rows, size_t cols)
 	RowSet* pRow = malloc(sizeof(RowSet));
 	assert(pRow);
 
-	pRow->widths = (size_t*)malloc(cols * sizeof(int));
-	memset(pRow->widths, 0, cols * sizeof(int));
+	pRow->widths = (size_t*)malloc(cols * sizeof(size_t));
+	memset(pRow->widths, 0, cols * sizeof(size_t));
 
 	pRow->data = malloc(rows * sizeof(char*));
 	for (i = 0; i < rows; i++)
